abke/kklmr16: Simplified Decrypt() interface.
The Decrypt() interfaces in the abke C code allow
the caller to specify attributes that are decrypted, but the
only exmaples that I found are where the attributes are teh same
as those used to generate the correcponding private key.  Thus,
to simplify the interface, I removed this parameter to
Decrypt and added an Attrs field to the PrivateKey struct that
stores the attributes corresponding to that key.

In short, this simplifies the interface, making it less
confusing.